Consolidate YAML objects into inner classes. Rename classes to match spec.
